New gitignore to skip system files
Updated various IDs in the models from int to long
Added Status property in PeriodType
Added ParentId property in FixturesEvent
New enum with the various period status
Added optional baseAddress to PinnacleClient constructor (pinnacle shifted some of its business under alternative URLs)
Added optional apiVersion parameter in GetOddsRequest and GetFixturesRequests because the version isn't the same between the various domains (some use v3/odds, others v1/odds...)

+++ Data/PlaceBetRequest.cs
@@ -46,7 +46,7 @@ namespace PinnacleWrapper.Data
         public int LineId;
 
         [JsonProperty(PropertyName = "altLineId")]      // Alternate line identification. Not required
-        public int? AltLineId;
+        public long? AltLineId;
 
         [JsonProperty(PropertyName = "pitcher1MustStart")] // Baseball only. Refers to the pitcher for TEAM_TYPE.Team1. This applicable only for MONEYLINE bet type, for all other bet types this has to be TRUE
         public bool? Pitcher1MustStart;
